separate ObsActions into its own object
