+++ explore/src/main/scala/explore/observationtree/package.scala
@@ -11,7 +11,6 @@ import crystal.react.*
 import eu.timepit.refined.types.numeric.NonNegInt
 import explore.DefaultErrorPolicy
 import explore.common.GroupQueries
-import explore.data.KeyedIndexedList
 import explore.data.tree.KeyedIndexedTree.Index
 import explore.data.tree.Node
 import explore.model.AppContext
@@ -23,7 +22,6 @@ import explore.model.enums.AppTab
 import explore.optics.GetAdjust
 import explore.optics.all.*
 import explore.syntax.ui.*
-import explore.undo.Action
 import explore.undo.KIListMod
 import explore.undo.KITreeMod
 import explore.undo.UndoSetter
@@ -72,120 +70,32 @@ def cloneObs(
   before >>
     cloneObservation[IO](obsId, newGroupId)
       .flatMap: newObs =>
-        obsExistence(newObs.id, o => setObs(programId, o.some, ctx))
+        ObsActions
+          .obsExistence(newObs.id, o => setObs(programId, o.some, ctx))
           .mod(observations):
             // Just place the new obs at the end of the group, which is where the server clones it.
             obsListMod.upsert(newObs, NonNegInt.MaxValue)
           .toAsync
       .guarantee(after)
 
-private def obsWithId(
-  obsId: Observation.Id
-): GetAdjust[KeyedIndexedList[Observation.Id, Observation], Option[
-  Observation
-]] =
+private def obsWithId(obsId: Observation.Id): GetAdjust[ObservationList, Option[Observation]] =
   obsListMod
     .withKey(obsId)
     .composeOptionLens(Focus[(Observation, NonNegInt)](_._1))
 
-def obsEditStatus(obsId: Observation.Id)(using
-  FetchClient[IO, ObservationDB]
-) = Action(
-  access = obsWithId(obsId).composeOptionLens(Observation.status)
-)(onSet =
-  (_, status) =>
-    UpdateObservationMutation[IO]
-      .execute(
-        UpdateObservationsInput(
-          WHERE = obsId.toWhereObservation.assign,
-          SET = ObservationPropertiesInput(status = status.orIgnore)
-        )
-      )
-      .void
-)
-
-def obsEditSubtitle(obsId: Observation.Id)(using
-  FetchClient[IO, ObservationDB]
-) = Action(
-  access = obsWithId(obsId).composeOptionLens(Observation.subtitle)
-)(onSet =
-  (_, subtitleOpt) =>
-    UpdateObservationMutation[IO]
-      .execute(
-        UpdateObservationsInput(
-          WHERE = obsId.toWhereObservation.assign,
-          SET = ObservationPropertiesInput(subtitle = subtitleOpt.flatten.orUnassign)
-        )
-      )
-      .void
-)
 def obsEditAttachments(
   obsId:         Observation.Id,
   attachmentIds: Set[ObsAttachment.Id]
 )(using
   FetchClient[IO, ObservationDB]
-) =
+): IO[Unit] =
   UpdateObservationMutation[IO]
-    .execute(
+    .execute:
       UpdateObservationsInput(
         WHERE = obsId.toWhereObservation.assign,
         SET = ObservationPropertiesInput(obsAttachments = attachmentIds.toList.assign)
       )
-    )
     .void
-
-def obsActiveStatus(obsId: Observation.Id)(using
-  FetchClient[IO, ObservationDB]
-) = Action(
-  access = obsWithId(obsId).composeOptionLens(Observation.activeStatus)
-)(onSet =
-  (_, activeStatus) =>
-    UpdateObservationMutation[IO]
-      .execute(
-        UpdateObservationsInput(
-          WHERE = obsId.toWhereObservation.assign,
-          SET = ObservationPropertiesInput(activeStatus = activeStatus.orIgnore)
-        )
-      )
-      .void
-)
-
-def obsScienceBand(obsId: Observation.Id)(using FetchClient[IO, ObservationDB]) =
-  Action(
-    access = obsWithId(obsId).composeOptionOptionLens(Observation.scienceBand)
-  )(
-    onSet = (_, scienceBand) =>
-      UpdateObservationMutation[IO]
-        .execute(
-          UpdateObservationsInput(
-            WHERE = obsId.toWhereObservation.assign,
-            SET = ObservationPropertiesInput(scienceBand = scienceBand.orUnassign)
-          )
-        )
-        .void
-  )
-
-def obsExistence(obsId: Observation.Id, setObs: Observation.Id => Callback)(using
-  FetchClient[IO, ObservationDB]
-) =
-  Action(
-    access = obsListMod.withKey(obsId)
-  )(
-    onSet = (_, elemWithIndexOpt) =>
-      elemWithIndexOpt.fold {
-        deleteObservation[IO](obsId)
-      } { case (obs, _) =>
-        // Not much to do here, the observation must be created before we get here
-        setObs(obs.id).toAsync
-      },
-    onRestore = (_, elemWithIndexOpt) =>
-      elemWithIndexOpt.fold {
-        deleteObservation[IO](obsId)
-      } { case (obs, _) =>
-        undeleteObservation[IO](obs.id) >>
-          setObs(obs.id).toAsync
-      }
-  )
 
 object AddingObservation extends NewType[Boolean]
 type AddingObservation = AddingObservation.Type
@@ -203,7 +113,8 @@ def insertObs(
 
   createObservation[IO](programId, parentId)
     .flatMap: (obs, groupIndex) =>
-      (obsExistence(obs.id, o => setObs(programId, o.some, ctx))
+      (ObsActions
+        .obsExistence(obs.id, o => setObs(programId, o.some, ctx))
         .mod(observations)(obsListMod.upsert(obs, pos)) >>
         groupTree.mod:
           _.inserted(
@@ -219,33 +130,6 @@ private def findGrouping(
 ): GroupTree => Option[(GroupTree.Node, GroupTree.Index)] =
   _.getNodeAndIndexByKey(groupId.asRight)
 
-def groupExistence(
-  groupId:  Group.Id,
-  setGroup: Group.Id => Callback
-)(using
-  FetchClient[IO, ObservationDB]
-): Action[GroupTree, Option[(GroupTree.Node, GroupTree.Index)]] = Action(
-  getter = findGrouping(groupId),
-  setter = maybeGroup =>
-    groupList =>
-      maybeGroup match
-        case None              => groupList.removed(groupId.asRight)
-        case Some((node, idx)) =>
-          val group = node.value
-          groupList.upserted(groupId.asRight, group, idx)
-)(
-  onSet = (_, node) =>
-    node.fold {
-      GroupQueries.deleteGroup[IO](groupId)
-    } { case (_, _) => setGroup(groupId).toAsync },
-  onRestore = (_, node) =>
-    node.fold {
-      GroupQueries.deleteGroup[IO](groupId)
-    } { case (_, _) =>
-      GroupQueries.undeleteGroup[IO](groupId) >> setGroup(groupId).toAsync
-    }
-)
-
 def insertGroup(
   programId: Program.Id,
   parentId:  Option[Group.Id],
@@ -258,7 +142,8 @@ def insertGroup(
   GroupQueries
     .createGroup[IO](programId, parentId)
     .flatMap: (group, parentIndex) =>
-      groupExistence(group.id, g => setGroup(programId, g.some, ctx))
+      ObsActions
+        .groupExistence(group.id, g => setGroup(programId, g.some, ctx))
         .set(groups):
           (Node(ServerIndexed(group.asRight, parentIndex)),
            Index(parentId.map(_.asRight), NonNegInt.MaxValue)
